requirements.txt
# TacBench dependencies
# Requires Python >= 3.11

# StarCraft II bot framework (actively maintained python-sc2 fork)
burnysc2>=0.11.0

# Settings management via .env file
pydantic-settings>=2.0.0
